+++ ChillaxScraps/Utils/Effects.cs
@@ -228,6 +228,11 @@ namespace ChillaxScraps.Utils
         {
             var original = StartOfRound.Instance.currentLevel.currentWeather;
             StartOfRound.Instance.currentLevel.currentWeather = weather;
+            if (Plugin.config.WeatherRegistery)
+            {
+                ChangeWeatherWR(weather);
+                return;
+            }
             RoundManager.Instance.SetToCurrentLevelWeather();
             TimeOfDay.Instance.SetWeatherBasedOnVariables();
             if (GameNetworkManager.Instance.localPlayerController.isInsideFactory)
@@ -265,6 +270,12 @@ namespace ChillaxScraps.Utils
                 player.isMovementHindered = Mathf.Clamp(player.isMovementHindered - 1, 0, 100);
                 player.hinderedMultiplier = 1f;
             }
+        }
+
+        public static void ChangeWeatherWR(LevelWeatherType weather)
+        {
+            if (GameNetworkManager.Instance.localPlayerController.IsHost)
+                WeatherRegistry.WeatherController.SetWeatherEffects(weather);
         }
 
         public static NetworkObjectReference Spawn(SpawnableEnemyWithRarity enemy, Vector3 position, float yRot = 0f)